Validation

+++ controllers/comments.js
@@ -1,55 +1,58 @@
 const Comment = require("../models/comments")
-const {getPostById} = require('../controllers/post');
-const req = require("express/lib/request");
+const Post = require('../models/post');
+const AppError = require('../utils/AppError')
+
 
 const getComments = async (req, res, next) => {
   try{
     const { postId } = req.params;
-    const post = getPostById(postId);
-    const comments = await Comment.find({ post: postId})
-    if(comments.length() < 1){
-      res.status(404).json({
-        message: "No comments for this post yet"
-      })
+    const post = await Post.findById(postId);
+    if(!post){
+      throw new AppError("No post with that Id", 404)
     }
+    const comments = await Comment.find({ post: postId})
     res.status(200).json({
       comments
     })
   }catch(err){
-    console.error(err)
+    next(err)
   }
 }
 
 const createComment = async (req, res, next) => {
   try{
-    const {postId} = req.params;
+    const postId = req.params.postId;
     const { content, author } = req.body;
-    const post = getPostById(postId);
-    const comment = await Comment.create({post: postId, author: author, content })
+    const post = await Post.findById(postId);
+    if(!post){
+      throw new AppError("No post with that Id", 404)
+    }
+    const comment = await Comment.create({post: postId, author, content })
     res.status(201).json({
       message: "Comment added!",
       comment
     })
   }catch(err){
-    console.error(err)
+    next(err)
   }
 }
 
 const deleteComment = async (req, res, next) => {
   try{
     const {postId, id } = req.params;
-    const post = getPostById(postId);
+    const post = await Post.findById(postId);
+    if(!post){
+      throw new AppError("No post with that Id", 404)
+    }
     const comment = await Comment.findByIdAndDelete(id);
     if(!comment){
-      res.status(404).json({
-        message: "No comment with that id"
-      })
+      throw new AppError('No comment with that Id', 404)
     }
     res.status(200).json({
       message: "Comment deleted."
     })
   }catch(err){
-    console.error(err);
+    next(err)
   }
 }
